.azuredevops: Add initial Azure Devops pipeline
Adds a pipeline to bulid and test Uncrustify debug/release builds
on Ubuntu and Windows.

As-is, this does not support the complete set of functionality in
the upstream Uncrustify Appveyor and Travis CI builds but it does
cover build and cTest unit testing which are most critical for
ensuring stability and readiness for upstream changes.

This pipeline will verify build and test of new Uncrustify changes
picked up directly from the upstream repo and be used to test
additional changes made in this fork.