Merged PR 22: .azuredevops/NugetPulishingStage.yml: Publish a Debug and Release package
Updates the NuGet publishing code to publish two packages instead
of one - a Debug package and a Release package.

Each package has directories at the package root which specify the
OS and architecture that executable is built for. This allows a user
interested in the Release binaries to get both the Linux and Windows
binaries in a single download. This is particuarly useful in some
CI environments where the NuGet package might be pulled down onto
a Linux or Windows build agent.