Prevent errors in typedef function prototypes
Currently, function prototypes can be formatted in a way that is no
longer compilable. A qualifier, type, and following chunk such as
a pointer asterisk can be appended into a single string.

This change updates the conditions for a previous boolean expression
that used indent_func_call_edk2_style to prevent the chunks from
being appended in typedefs.