.github: Fix mislabelling of macos-latest arm64 build as x64
build-and-test.yml must have been building the macOS version
of Tianocore Uncrustify since a time when macos-latest invoked
an X64 runner. However, it now invokes an Apple ARM runner.
Nothing in 'Build (macOS)' is influenced in any way by the target
architecture - except for what it names the output file - so the CI
just goes ahead and builds a perfectly good MacOs-ARM version,
which it  wrongly publishes as a MacOs-x86 version.

There is believed to be no need to produce a MacOs-x86 version
any more (and probably no simple way to), so we just go ahead and
fix up naming so that future releases contain a MacOs-ARM version
labelled as such.

Fixes: #7

Signed-off-by: Mike Beaton <mjsbeaton@gmail.com>