.github: Switch to native arm64 python for macos-latest builds
On building the version of actions produced for #7 (to correctly reflect
the fact that macos-latest now produces an ARM build), it was noticed
(e.g. in 'Copy macOS Executable' step) that an x64 version of python
is being used in this build. While this shouldn't matter, it makes more
sense to use a native version.

Since Windows and Ubuntu arm64 builds run on x64 hosts, we cannot just
follow matrix.target_arch; in order to use arm64 python only on macOS
we need the slightly more complex logic which is applied here.

Signed-off-by: Mike Beaton <mjsbeaton@gmail.com>